[UPD] cg: move files writer code_generator
