+++ code_generator/models/code_generator_writer.py
@@ -12,14 +12,12 @@ from lxml.builder import E
 from lxml import etree as ET
 from collections import defaultdict
 from odoo.tools.misc import mute_logger
-import subprocess
-import html5print
-import xmlformatter
 from PIL import Image
 import io
 
-import xml.dom.minicompat
-from xml.dom import minidom, Node
+from ..extractor_module import ExtractorModule
+from ..extractor_view import ExtractorView
+from ..code_generator_data import CodeGeneratorData
 
 from code_writer import CodeWriter
 from odoo.models import MAGIC_COLUMNS
@@ -64,1275 +62,6 @@ class CodeGeneratorWriter(models.Model):
     rootdir = fields.Char(string="Root dir")
 
     basename = fields.Char(string="Base name")
-
-    class ExtractorView:
-        def __init__(self, module, model_model):
-            self._module = module
-            self.view_ids = module.env["ir.ui.view"].search(
-                [("model", "=", model_model)]
-            )
-            self.code_generator_id = None
-            self.model_id = module.env["ir.model"].search(
-                [("model", "=", model_model)]
-            )
-            self.dct_model = defaultdict(dict)
-            self.dct_field = defaultdict(dict)
-            self.module_attr = defaultdict(dict)
-            model_name = model_model.replace(".", "_")
-            self.var_model_name = f"model_{model_name}"
-            self.var_model = model_model
-            if self.view_ids:
-                # create temporary module
-                name = f"TEMP_{model_name}"
-                i = 1
-                while module.env["code.generator.module"].search(
-                    [("name", "=", name)]
-                ):
-                    name = f"TEMP_{i}_{model_name}"
-                    i += 1
-                value = {
-                    "name": name,
-                    "shortdesc": "None",
-                }
-                self.code_generator_id = module.env[
-                    "code.generator.module"
-                ].create(value)
-                self._parse_view_ids()
-                self._parse_menu()
-                self._parse_action_server()
-
-        def _parse_action_server(self):
-            # Search comment node associated to action_server
-            module = self._module
-            if not module.template_module_path_generated_extension:
-                return
-            relative_path_generated_module = (
-                module.template_module_path_generated_extension.replace(
-                    "'", ""
-                ).replace(", ", "/")
-            )
-            path_generated_module = os.path.normpath(
-                os.path.join(
-                    module.path_sync_code,
-                    relative_path_generated_module,
-                    module.template_module_name,
-                    "**",
-                    "*.xml",
-                )
-            )
-            lst_xml_file = glob.glob(path_generated_module)
-            for xml_file in lst_xml_file:
-                my_xml = minidom.parse(xml_file)
-                lst_record = my_xml.getElementsByTagName("record")
-                for record in lst_record:
-                    # detect action_server_backup
-                    searched_record = "model", "ir.actions.server"
-
-                    if searched_record in record.attributes.items():
-                        last_record = record.previousSibling.previousSibling
-                        if last_record.nodeType is Node.COMMENT_NODE:
-                            record_id = dict(record.attributes.items()).get(
-                                "id"
-                            )
-                            if not record_id:
-                                _logger.warning(
-                                    "Missing id when searching"
-                                    " ir.actions.server."
-                                )
-                                continue
-                            xml_id = (
-                                f"{module.template_module_name}.{record_id}"
-                            )
-                            result = self._module.env.ref(
-                                xml_id, raise_if_not_found=False
-                            )
-                            if result:
-                                result.comment = last_record.data.strip()
-
-        def _parse_menu(self):
-            ir_model_data_ids = self._module.env["ir.model.data"].search(
-                [
-                    ("model", "=", "ir.ui.menu"),
-                    ("module", "=", self._module.template_module_name),
-                ]
-            )
-            if not ir_model_data_ids:
-                return
-            lst_id_menu = [a.res_id for a in ir_model_data_ids]
-            menu_ids = self._module.env["ir.ui.menu"].browse(lst_id_menu)
-            for menu_id in menu_ids:
-
-                # TODO optimise request ir.model.data, this is duplicated
-                menu_action = None
-                if menu_id.action:
-                    # Create act_window
-                    menu_data_id = self._module.env["ir.model.data"].search(
-                        [
-                            ("model", "=", "ir.actions.act_window"),
-                            ("res_id", "=", menu_id.action.id),
-                        ]
-                    )
-                    dct_act_value = {
-                        "id_name": menu_data_id.name,
-                        "name": menu_id.action.name,
-                        "code_generator_id": self.code_generator_id.id,
-                    }
-                    menu_action = self._module.env[
-                        "code.generator.act_window"
-                    ].create(dct_act_value)
-                # Create menu
-                menu_data_id = self._module.env["ir.model.data"].search(
-                    [("model", "=", "ir.ui.menu"), ("res_id", "=", menu_id.id)]
-                )
-                dct_menu_value = {
-                    "code_generator_id": self.code_generator_id.id,
-                    "id_name": menu_data_id.name,
-                }
-                if menu_id.sequence != 10:
-                    dct_menu_value["sequence"] = menu_id.sequence
-                if menu_id.parent_id:
-                    menu_data_parent_id = self._module.env[
-                        "ir.model.data"
-                    ].search(
-                        [
-                            ("model", "=", "ir.ui.menu"),
-                            ("res_id", "=", menu_id.parent_id.id),
-                        ]
-                    )
-                    dct_menu_value[
-                        "parent_id_name"
-                    ] = menu_data_parent_id.complete_name
-
-                if menu_action:
-                    dct_menu_value["m2o_act_window"] = menu_action.id
-
-                self._module.env["code.generator.menu"].create(dct_menu_value)
-                # If need to associated
-                # menu_id.m2o_module = self._module.id
-
-        def _parse_view_ids(self):
-            for view_id in self.view_ids:
-                mydoc = minidom.parseString(view_id.arch_base.encode())
-
-                lst_view_item_id = []
-
-                # Search form
-                lst_form_xml = mydoc.getElementsByTagName("form")
-                if lst_form_xml:
-                    sequence_form = 10
-                    lst_form_field_xml = mydoc.getElementsByTagName("field")
-                    for field_xml in lst_form_field_xml:
-                        field_name = dict(field_xml.attributes.items()).get(
-                            "name"
-                        )
-                        if field_name in self.dct_model[view_id.model]:
-                            self.dct_model[view_id.model][field_name][
-                                "code_generator_form_simple_view_sequence"
-                            ] = sequence_form
-                        else:
-                            self.dct_model[view_id.model][field_name] = {
-                                "code_generator_form_simple_view_sequence": sequence_form
-                            }
-                        sequence_form += 1
-
-                # Search tree
-                lst_tree_xml = mydoc.getElementsByTagName("tree")
-                if lst_tree_xml:
-                    sequence_tree = 10
-                    lst_tree_field_xml = mydoc.getElementsByTagName("field")
-                    for field_xml in lst_tree_field_xml:
-                        field_name = dict(field_xml.attributes.items()).get(
-                            "name"
-                        )
-                        if field_name in self.dct_model[view_id.model]:
-                            self.dct_model[view_id.model][field_name][
-                                "code_generator_tree_view_sequence"
-                            ] = sequence_tree
-                        else:
-                            self.dct_model[view_id.model][field_name] = {
-                                "code_generator_tree_view_sequence": sequence_tree
-                            }
-                        sequence_tree += 1
-
-                # Search timeline
-                lst_timeline_xml = mydoc.getElementsByTagName("timeline")
-                if lst_timeline_xml:
-                    for timeline_xml in lst_timeline_xml:
-                        if "date_start" in timeline_xml.attributes.keys():
-                            field_name = (
-                                timeline_xml.attributes["date_start"]
-                                .childNodes[0]
-                                .data
-                            )
-                            for field_id in self.model_id.field_id:
-                                if field_id.name == field_name:
-                                    self.dct_field[field_name][
-                                        "is_date_start_view"
-                                    ] = True
-                        if "date_stop" in timeline_xml.attributes.keys():
-                            field_name = (
-                                timeline_xml.attributes["date_stop"]
-                                .childNodes[0]
-                                .data
-                            )
-                            for field_id in self.model_id.field_id:
-                                if field_id.name == field_name:
-                                    self.dct_field[field_name][
-                                        "is_date_end_view"
-                                    ] = True
-
-                # Search diagram
-                lst_diagram_xml = mydoc.getElementsByTagName("diagram")
-                if lst_diagram_xml:
-                    nb_iter = 0
-                    for diagram_xml in lst_diagram_xml:
-                        nb_iter += 1
-                        if nb_iter > 1:
-                            _logger.warning(
-                                "Cannot support multiple diagram in the same"
-                                " file."
-                            )
-                            continue
-                        # Search 1 node, 1 arrow and maybe 1 label
-                        find_node = False
-                        find_arrow = False
-                        find_label = False
-                        for child_div in diagram_xml.childNodes:
-                            if child_div.nodeType is Node.ELEMENT_NODE:
-                                dct_att = dict(child_div.attributes.items())
-                                if child_div.nodeName == "node":
-                                    find_node = True
-                                    node_object = dct_att.get("object")
-                                    node_xpos = dct_att.get("xpos")
-                                    node_ypos = dct_att.get("ypos")
-                                    node_shape = dct_att.get("shape")
-                                    node_form_view_ref = dct_att.get(
-                                        "form_view_ref"
-                                    )
-                                    if not node_object:
-                                        _logger.warning(
-                                            "Missing diagram node object for"
-                                            f" model {self.var_model}"
-                                        )
-                                        find_node = False
-                                    if not node_xpos:
-                                        _logger.warning(
-                                            "Missing diagram node xpos for"
-                                            f" model {self.var_model}"
-                                        )
-                                        find_node = False
-                                    if not node_ypos:
-                                        _logger.warning(
-                                            "Missing diagram node ypos for"
-                                            f" model {self.var_model}"
-                                        )
-                                        find_node = False
-                                    if not node_shape:
-                                        _logger.warning(
-                                            "Missing diagram node shape for"
-                                            f" model {self.var_model}"
-                                        )
-                                        find_node = False
-                                    if not node_form_view_ref:
-                                        _logger.warning(
-                                            "Missing diagram node"
-                                            " form_view_ref for model"
-                                            f" {self.var_model}"
-                                        )
-                                        find_node = False
-                                if child_div.nodeName == "arrow":
-                                    find_arrow = True
-                                    arrow_object = dct_att.get("object")
-                                    arrow_source = dct_att.get("source")
-                                    arrow_destination = dct_att.get(
-                                        "destination"
-                                    )
-                                    arrow_label = dct_att.get("label")
-                                    arrow_form_view_ref = dct_att.get(
-                                        "form_view_ref"
-                                    )
-                                    if not arrow_object:
-                                        _logger.warning(
-                                            "Missing diagram arrow object for"
-                                            f" model {self.var_model}"
-                                        )
-                                        find_arrow = False
-                                    if not arrow_source:
-                                        _logger.warning(
-                                            "Missing diagram arrow source for"
-                                            f" model {self.var_model}"
-                                        )
-                                        find_arrow = False
-                                    if not arrow_destination:
-                                        _logger.warning(
-                                            "Missing diagram arrow"
-                                            " destination for model"
-                                            f" {self.var_model}"
-                                        )
-                                        find_arrow = False
-                                    if not arrow_label:
-                                        _logger.warning(
-                                            "Missing diagram arrow label for"
-                                            f" model {self.var_model}"
-                                        )
-                                        find_arrow = False
-                                    if not arrow_form_view_ref:
-                                        _logger.warning(
-                                            "Missing diagram arrow"
-                                            " form_view_ref for model"
-                                            f" {self.var_model}"
-                                        )
-                                        find_arrow = False
-                                if child_div.nodeName == "label":
-                                    find_label = True
-                                    diagram_label_string = dct_att.get(
-                                        "string"
-                                    )
-                                    if not diagram_label_string:
-                                        _logger.warning(
-                                            "Missing diagram label string"
-                                            f" for model {self.var_model}"
-                                        )
-                                        find_label = False
-                        if find_node and find_arrow:
-                            self.model_id.diagram_node_object = node_object
-                            self.model_id.diagram_node_xpos_field = node_xpos
-                            self.model_id.diagram_node_ypos_field = node_ypos
-                            self.model_id.diagram_node_shape_field = node_shape
-                            self.model_id.diagram_node_form_view_ref = (
-                                node_form_view_ref
-                            )
-                            self.model_id.diagram_arrow_object = arrow_object
-                            self.model_id.diagram_arrow_src_field = (
-                                arrow_source
-                            )
-                            self.model_id.diagram_arrow_dst_field = (
-                                arrow_destination
-                            )
-                            if find_label:
-                                self.model_id.diagram_label_string = (
-                                    diagram_label_string
-                                )
-
-                # Search oe_chatter activity message_ids or message_follower_ids
-                lst_div_xml = mydoc.getElementsByTagName("div")
-                if lst_div_xml:
-                    for div_xml in lst_div_xml:
-                        if (
-                            "class",
-                            "oe_chatter",
-                        ) in div_xml.attributes.items():
-                            for child_div in div_xml.childNodes:
-                                if child_div.nodeType is Node.ELEMENT_NODE:
-                                    lst_value = dict(
-                                        child_div.attributes.items()
-                                    ).values()
-                                    if (
-                                        "activity_ids" in lst_value
-                                        or "message_ids" in lst_value
-                                        or "message_follower_ids" in lst_value
-                                    ):
-                                        # self.model_id.write(
-                                        #     {"enable_activity": True}
-                                        # )
-                                        self.model_id.enable_activity = True
-
-                # Sheet
-                lst_sheet_xml = mydoc.getElementsByTagName("sheet")
-                has_body_sheet = bool(lst_sheet_xml)
-                sheet_xml = lst_sheet_xml[0] if lst_sheet_xml else None
-                if len(lst_sheet_xml) > 1:
-                    _logger.warning("Cannot support multiple <sheet>.")
-
-                # Search header
-                header_xml = None
-                no_sequence = 1
-                lst_header_xml = mydoc.getElementsByTagName("header")
-                if len(lst_header_xml) > 1:
-                    _logger.warning("Cannot support multiple header.")
-                for header_xml in lst_header_xml:
-                    # TODO get inside attributes for header
-                    for child_header in header_xml.childNodes:
-                        if child_header.nodeType is Node.TEXT_NODE:
-                            data = child_header.data.strip()
-                            if data:
-                                _logger.warning("Not supported.")
-                        elif child_header.nodeType is Node.ELEMENT_NODE:
-                            self._extract_child_xml(
-                                child_header,
-                                lst_view_item_id,
-                                "header",
-                                sequence=no_sequence,
-                            )
-
-                # Search title
-                no_sequence = 1
-                nb_oe_title = 0
-                div_title = None
-                for div_xml in mydoc.getElementsByTagName("div"):
-                    # Find oe_title class
-                    # TODO what todo when multiple class? split by ,
-                    for key, value in div_xml.attributes.items():
-                        if key == "class" and value == "oe_title":
-                            div_title = div_xml
-                            nb_oe_title += 1
-                            if nb_oe_title > 1:
-                                _logger.warning(
-                                    "Cannot support multiple class oe_title."
-                                )
-                                continue
-                            # TODO support multiple element in title
-                            lst_field = div_xml.getElementsByTagName("field")
-                            if not lst_field:
-                                _logger.warning(
-                                    "Not supported title without field, TODO."
-                                )
-                            elif len(lst_field) > 1:
-                                _logger.warning(
-                                    "Not supported title without multiple"
-                                    " field, TODO."
-                                )
-                            else:
-                                dct_field_attrs = dict(
-                                    div_xml.getElementsByTagName("field")[
-                                        0
-                                    ].attributes.items()
-                                )
-                                name = dct_field_attrs.get("name")
-                                if not name:
-                                    _logger.warning(
-                                        "Cannot identify field type in title."
-                                    )
-                                else:
-                                    dct_attributes = {
-                                        "action_name": name,
-                                        "section_type": "title",
-                                        "item_type": "field",
-                                        "sequence": no_sequence,
-                                    }
-                                    view_item_id = self._module.env[
-                                        "code.generator.view.item"
-                                    ].create(dct_attributes)
-                                    lst_view_item_id.append(view_item_id.id)
-                                    no_sequence += 1
-
-                lst_body_xml = []
-                lst_tag_support = list(
-                    dict(
-                        self._module.env["code.generator.view"]
-                        ._fields["view_type"]
-                        .selection
-                    ).keys()
-                )
-                lst_content = [
-                    b
-                    for a in lst_tag_support
-                    for b in mydoc.getElementsByTagName(a)
-                ]
-                if not lst_content:
-                    _logger.warning(
-                        f"Cannot find a xml type from list: {lst_tag_support}."
-                    )
-                elif len(lst_content) > 1:
-                    _logger.warning(
-                        f"Cannot support multiple {lst_tag_support}."
-                    )
-                else:
-                    form_xml = lst_content[0]
-                    for child_form in form_xml.childNodes:
-                        if child_form.nodeType is Node.TEXT_NODE:
-                            data = child_form.data.strip()
-                            if data:
-                                _logger.warning("Not supported.")
-                        elif child_form.nodeType is Node.ELEMENT_NODE:
-                            if (
-                                child_form == div_title
-                                or child_form == header_xml
-                                or child_form == sheet_xml
-                            ):
-                                continue
-                            # if has_body_sheet:
-                            #     _logger.warning(
-                            #         "How can find body xml outside of his"
-                            #         " sheet?"
-                            #     )
-                            # else:
-                            #     lst_body_xml.append(child_form)
-                            # TODO everything can be in body? check when add oe_chatter
-                            lst_body_xml.append(child_form)
-
-                if lst_sheet_xml:
-                    # TODO validate this, test with and without <sheet>
-                    if type(lst_sheet_xml) is xml.dom.minicompat.NodeList:
-                        lst_body_xml = [a for a in lst_sheet_xml[0].childNodes]
-                    else:
-                        lst_body_xml = [a for a in lst_sheet_xml.childNodes]
-                sequence = 1
-                lst_node = []
-                for body_xml in lst_body_xml:
-                    if body_xml.nodeType is Node.TEXT_NODE:
-                        data = body_xml.data.strip()
-                        if data:
-                            _logger.warning(f"Not supported : {data}.")
-                    elif body_xml.nodeType is Node.ELEMENT_NODE:
-                        status = self._extract_child_xml(
-                            body_xml,
-                            lst_view_item_id,
-                            "body",
-                            lst_node=lst_node,
-                            sequence=sequence,
-                        )
-                        if status:
-                            lst_node.append(body_xml)
-                        else:
-                            lst_node = []
-                        sequence += 1
-                if lst_node:
-                    _logger.warning("Missing node in buffer.")
-
-                value = {
-                    "code_generator_id": self.code_generator_id.id,
-                    "view_type": view_id.type,
-                    # "view_name": "view_backup_conf_form",
-                    # "m2o_model": model_db_backup.id,
-                    "view_item_ids": [(6, 0, lst_view_item_id)],
-                    "has_body_sheet": has_body_sheet,
-                }
-
-                # ID
-                ir_model_data = self._module.env["ir.model.data"].search(
-                    [
-                        ("model", "=", "ir.ui.view"),
-                        ("res_id", "=", view_id.id),
-                    ]
-                )
-                if ir_model_data:
-                    first_name = ir_model_data[0].name
-                    if len(ir_model_data) > 1:
-                        _logger.warning(
-                            f"Duplicated view model id {first_name}"
-                        )
-                    value["id_name"] = first_name
-                view_code_generator = self._module.env[
-                    "code.generator.view"
-                ].create(value)
-
-        def _extract_child_xml(
-            self,
-            node,
-            lst_view_item_id,
-            section_type,
-            lst_node=[],
-            parent=None,
-            sequence=1,
-        ):
-            """
-
-            :param node:
-            :param lst_view_item_id:
-            :param section_type:
-            :param parent:
-            :param sequence:
-            :return: when True, cumulate the node in lst_node for next run, else None
-            """
-            # From background_type
-            lst_key_html_class = (
-                "bg-success",
-                "bg-success-full",
-                "bg-warning",
-                "bg-warning-full",
-                "bg-info",
-                "bg-info-full",
-                "bg-danger",
-                "bg-danger-full",
-                "bg-light",
-                "bg-dark",
-            )
-            dct_key_keep = {
-                "name": "action_name",
-                "string": "label",
-                "attrs": "attrs",
-            }
-            dct_attributes = {
-                "section_type": section_type,
-                "item_type": node.nodeName,
-                "sequence": sequence,
-            }
-
-            if parent:
-                dct_attributes["parent_id"] = parent.id
-
-            if node.nodeName in (
-                "group",
-                "div",
-                "templates",
-                "t",
-                "ul",
-                "li",
-                "strong",
-                "i",
-            ):
-                if lst_node:
-                    # Check cached of nodes
-                    # maybe help node
-                    for cached_node in lst_node:
-                        # TODO need to check nodeName == "separator" ?
-                        for key, value in cached_node.attributes.items():
-                            if key == "string" and value == "Help":
-                                dct_attributes["is_help"] = True
-                            elif key == "colspan" and value != 1:
-                                dct_attributes["colspan"] = value
-                    dct_attributes["label"] = "\n".join(
-                        [a.strip() for a in node.toxml().split("\n")[1:-1]]
-                    )
-                    dct_attributes["item_type"] = "html"
-                else:
-                    for key, value in node.attributes.items():
-                        if key == "class" and value in lst_key_html_class:
-                            # not a real div, it's an html part
-                            dct_attributes["item_type"] = "html"
-                            dct_attributes["background_type"] = value
-                            text_html = ""
-                            for child in node.childNodes:
-                                # ignore element, only get text
-                                if child.nodeType is Node.TEXT_NODE:
-                                    data = child.data.strip()
-                                    if data:
-                                        text_html += data
-                                elif child.nodeType is Node.ELEMENT_NODE:
-                                    continue
-                            dct_attributes["label"] = text_html
-
-            elif node.nodeName == "button":
-                dct_key_keep["class"] = "button_type"
-                for key, value in node.attributes.items():
-                    if key == "icon":
-                        dct_attributes["icon"] = value
-            elif node.nodeName in ("field", "filter"):
-                for key, value in node.attributes.items():
-                    if key == "password":
-                        dct_attributes["password"] = value
-                    if key == "widget":
-                        field_name = dict(node.attributes.items()).get("name")
-                        # TODO update dict instead of overwrite it
-                        self.module_attr[self.var_model][field_name] = {
-                            "force_widget": value
-                        }
-                    if key == "placeholder":
-                        dct_attributes["placeholder"] = value
-            elif node.nodeName == "separator":
-                # Accumulate nodes
-                return True
-            elif node.nodeName == "templates":
-                _logger.warning(f"Node template is not supported, ignore it.")
-                return
-            else:
-                _logger.warning(f"Unknown this case '{node.nodeName}'.")
-                return
-
-            # TODO use external function to get attributes items to remove duplicate code, search "node.attributes.items()"
-            for key, value in node.attributes.items():
-                attributes_name = dct_key_keep.get(key)
-                if attributes_name:
-                    dct_attributes[attributes_name] = value
-            # TODO validate dct_attributes has all needed key with dct_key_keep (except button_type)
-            view_item_id = self._module.env["code.generator.view.item"].create(
-                dct_attributes
-            )
-            lst_view_item_id.append(view_item_id.id)
-            sequence += 1
-
-            # Child, except HTML
-            if dct_attributes["item_type"] != "html":
-                child_sequence = 1
-                for child in node.childNodes:
-                    if child.nodeType is Node.TEXT_NODE:
-                        data = child.data.strip()
-                        if data:
-                            _logger.warning(f"Not supported : {data}.")
-                    elif child.nodeType is Node.ELEMENT_NODE:
-                        self._extract_child_xml(
-                            child,
-                            lst_view_item_id,
-                            section_type,
-                            parent=view_item_id,
-                            sequence=child_sequence,
-                        )
-                        child_sequence += 1
-
-    class ExtractorModule:
-        def __init__(self, module, model_model, view_file_sync_model):
-            self.is_enabled = False
-            self.working_directory = module.path_sync_code
-            self.view_file_sync_model = view_file_sync_model
-            self.module = module
-            self.model = model_model
-            self.model_id = module.env["ir.model"].search(
-                [("model", "=", model_model)], limit=1
-            )
-            self.dct_model = view_file_sync_model.dct_model
-            self.py_filename = ""
-            if not module.template_module_path_generated_extension:
-                _logger.warning(
-                    f"The variable template_module_path_generated_extension is"
-                    f" empty."
-                )
-                return
-            if not self.model_id:
-                _logger.warning(f"Cannot found module {model_model}.")
-                return
-
-            relative_path_generated_module = (
-                module.template_module_path_generated_extension.replace(
-                    "'", ""
-                ).replace(", ", "/")
-            )
-            template_directory = os.path.normpath(
-                os.path.join(
-                    module.path_sync_code,
-                    relative_path_generated_module,
-                    module.template_module_name,
-                )
-            )
-            manifest_file_path = os.path.normpath(
-                os.path.join(
-                    template_directory,
-                    "__manifest__.py",
-                )
-            )
-
-            if module.template_module_id and os.path.isfile(
-                manifest_file_path
-            ):
-                with open(manifest_file_path, "r") as source:
-                    lst_line = source.readlines()
-                    i = 0
-                    for line in lst_line:
-                        if line.startswith("{"):
-                            break
-                        i += 1
-                str_line = "".join(lst_line[:i]).strip()
-                module.template_module_id.header_manifest = str_line
-                dct_data = ast.literal_eval("".join(lst_line[i:]).strip())
-                external_dep = dct_data.get("external_dependencies")
-                if external_dep:
-                    if type(external_dep) is dict:
-                        for key, lst_value in external_dep.items():
-                            if type(lst_value) is list:
-                                for value in lst_value:
-                                    v = {
-                                        "module_id": module.id,
-                                        "depend": value,
-                                        "application_type": key,
-                                        "is_template": True,
-                                    }
-                                    self.module.env[
-                                        "code.generator.module.external.dependency"
-                                    ].create(v)
-                            else:
-                                _logger.warning(
-                                    "Unknown value type external_dependencies"
-                                    f" in __manifest__ key {key}, value"
-                                    f" {value}."
-                                )
-                    else:
-                        _logger.warning(
-                            "Unknown external_dependencies in __manifest__"
-                            f" {external_dep}"
-                        )
-
-            elif not module.template_module_id:
-                _logger.warning(
-                    "Missing template_module_id in module to extract"
-                    " information."
-                )
-            elif not os.path.isfile(manifest_file_path):
-                _logger.warning(
-                    "Missing __manifest__.py file in directory"
-                    f" '{template_directory}' to extract information."
-                )
-
-            path_generated_module = os.path.normpath(
-                os.path.join(
-                    module.path_sync_code,
-                    relative_path_generated_module,
-                    module.template_module_name,
-                    "**",
-                    "*.py",
-                )
-            )
-            lst_py_file = glob.glob(path_generated_module)
-            if not lst_py_file:
-                return
-            for py_file in lst_py_file:
-                filename = py_file.split("/")[-1]
-                if filename == "__init__.py":
-                    continue
-                with open(py_file, "r") as source:
-                    f_lines = source.read()
-                    lst_lines = f_lines.split("\n")
-                    f_ast = ast.parse(f_lines)
-                    class_model_ast = self.search_class_model(f_ast)
-                    if class_model_ast:
-                        self.py_filename = filename
-                        self.search_field(class_model_ast)
-                        # Fill method
-                        self.search_import(lst_lines)
-                        self.search_method(class_model_ast, lst_lines, module)
-            self.is_enabled = True
-
-        def search_class_model(self, f_ast):
-            for children in f_ast.body:
-                # TODO check bases of class if equal models.Model for better performance
-                if type(children) == ast.ClassDef:
-                    # Detect good _name
-                    for node in children.body:
-                        if (
-                            type(node) is ast.Assign
-                            and node.targets
-                            and node.targets[0].id == "_name"
-                            and node.value.s == self.model
-                        ):
-                            return children
-
-        def extract_lambda(self, node):
-            args = ", ".join([a.arg for a in node.args.args])
-            value = ""
-            if type(node.body) is ast.Call:
-                # Support -> lambda self: self._default_folder()
-                body = node.body.func
-                value = f"{body.value.id}.{body.attr}()"
-            else:
-                _logger.error("Lambda not supported.")
-            return f"lambda {args}: {value}"
-
-        def _fill_search_field(self, ast_obj, var_name=""):
-            ast_obj_type = type(ast_obj)
-            if ast_obj_type is ast.Str:
-                result = ast_obj.s
-            elif ast_obj_type is ast.Lambda:
-                result = self.extract_lambda(ast_obj)
-            elif ast_obj_type is ast.NameConstant:
-                result = ast_obj.value
-            elif ast_obj_type is ast.Num:
-                result = ast_obj.n
-            elif ast_obj_type is ast.List:
-                lst_value = [
-                    self._fill_search_field(a, var_name) for a in ast_obj.elts
-                ]
-                result = lst_value
-            elif ast_obj_type is ast.Tuple:
-                lst_value = [
-                    self._fill_search_field(a, var_name) for a in ast_obj.elts
-                ]
-                result = tuple(lst_value)
-            else:
-                # TODO missing ast.Dict?
-                result = None
-                _logger.error(
-                    f"Cannot support keyword of variable {var_name} type"
-                    f" {ast_obj_type} in filename {self.py_filename}."
-                )
-            return result
-
-        def search_field(self, class_model_ast):
-            if self.dct_model[self.model]:
-                dct_field = self.dct_model[self.model]
-            else:
-                dct_field = {}
-                self.dct_model[self.model] = dct_field
-            lst_var_name_check = []
-
-            sequence = -1
-            for node in class_model_ast.body:
-                sequence += 1
-                if (
-                    type(node) is ast.Assign
-                    and type(node.value) is ast.Call
-                    and node.value.func.value.id == "fields"
-                ):
-                    var_name = node.targets[0].id
-                    d = {
-                        "type": node.value.func.attr,
-                        "code_generator_sequence": sequence,
-                    }
-                    for keyword in node.value.keywords:
-                        value = self._fill_search_field(
-                            keyword.value, var_name
-                        )
-                        # Waste to stock None value
-                        if value is not None:
-                            d[keyword.arg] = value
-                    if (
-                        self.view_file_sync_model
-                        and self.view_file_sync_model.module_attr
-                    ):
-                        dct_module_attr = (
-                            self.view_file_sync_model.module_attr.get(
-                                self.model
-                            )
-                        )
-                        if dct_module_attr:
-                            dct_field_module_attr = dct_module_attr.get(
-                                var_name
-                            )
-                            if dct_field_module_attr:
-                                for (
-                                    attr_key,
-                                    attr_value,
-                                ) in dct_field_module_attr.items():
-                                    d[attr_key] = attr_value
-
-                    if var_name in dct_field:
-                        dct_field[var_name].update(d)
-                    else:
-                        dct_field[var_name] = d
-                    lst_var_name_check.append(var_name)
-            # Remove item not from this list
-            lst_var_name_to_delete = list(
-                set(dct_field.keys()).difference(set(lst_var_name_check))
-            )
-            for var_name_to_delete in lst_var_name_to_delete:
-                del dct_field[var_name_to_delete]
-
-        def _extract_decorator(self, decorator_list):
-            str_decorator = ""
-            for dec in decorator_list:
-                if type(dec) is ast.Attribute:
-                    v = f"@{dec.value.id}.{dec.attr}"
-                elif type(dec) is ast.Call:
-                    args = [
-                        f'\\"{self._fill_search_field(a)}\\"' for a in dec.args
-                    ]
-                    str_arg = ", ".join(args)
-                    v = f"@{dec.func.value.id}.{dec.func.attr}({str_arg})"
-                elif type(dec) is ast.Name:
-                    v = f"@{dec.id}"
-                else:
-                    _logger.warning(
-                        f"Decorator type {type(dec)} not supported."
-                    )
-                    v = None
-
-                if v:
-                    if str_decorator:
-                        str_decorator += f";{v}"
-                    else:
-                        str_decorator = v
-            return str_decorator
-
-        def _write_exact_argument(self, value):
-            str_args = ""
-            if type(value) is ast.arg:
-                if hasattr(value, "is_vararg") and value.is_vararg:
-                    str_args += "*"
-                if hasattr(value, "is_kwarg") and value.is_kwarg:
-                    str_args += "**"
-                str_args += value.arg
-                if value.annotation:
-                    str_args += f": {value.annotation.id}"
-            else:
-                v = self._fill_search_field(value)
-                if type(v) is str:
-                    str_args += f"='{v}'"
-                else:
-                    str_args += f"={v}"
-            return str_args
-
-        def _extract_argument(self, ast_argument):
-            dct_args = {}
-            # Need to regroup different element in order
-            # Create dict with all element
-            if ast_argument.args:
-                for arg in ast_argument.args:
-                    dct_args[f"{arg.lineno}-{arg.col_offset}"] = arg
-            if ast_argument.defaults:
-                for arg in ast_argument.defaults:
-                    dct_args[f"{arg.lineno}-{arg.col_offset}"] = arg
-            if ast_argument.kwonlyargs:
-                for arg in ast_argument.kwonlyargs:
-                    dct_args[f"{arg.lineno}-{arg.col_offset}"] = arg
-            if ast_argument.kw_defaults:
-                for arg in ast_argument.kw_defaults:
-                    dct_args[f"{arg.lineno}-{arg.col_offset}"] = arg
-            if ast_argument.vararg:
-                arg = ast_argument.vararg
-                arg.is_vararg = True
-                dct_args[f"{arg.lineno}-{arg.col_offset}"] = arg
-            if ast_argument.kwarg:
-                arg = ast_argument.kwarg
-                arg.is_kwarg = True
-                dct_args[f"{arg.lineno}-{arg.col_offset}"] = arg
-
-            # Regroup all extra associated with arg
-            str_args = ""
-            lst_key_sorted = sorted(dct_args.keys())
-            lst_group_arg = []
-            last_lst_item = []
-            for key in lst_key_sorted:
-                value = dct_args[key]
-                if type(value) is ast.arg:
-                    # new item
-                    last_lst_item = [value]
-                    lst_group_arg.append(last_lst_item)
-                else:
-                    last_lst_item.append(value)
-
-            # Recreate string of argument
-            for lst_value in lst_group_arg[:-1]:
-                for value in lst_value:
-                    str_args += self._write_exact_argument(value)
-                str_args += ", "
-            last_value = lst_group_arg[-1]
-            if last_value:
-                for value in last_value:
-                    str_args += self._write_exact_argument(value)
-            return str_args
-
-        def _get_nb_line_multiple_string(
-            self, item, lst_line, i_lineno, extra_size=2
-        ):
-            str_size = len(item.s)
-            line_size = len(lst_line[i_lineno - 1].strip())
-            if line_size != str_size + extra_size:
-                # Try detect multiline string with pending technique like
-                # """test1"""
-                # """test2"""
-                # This will be """test1test2"""
-                # or
-                # "test1"
-                # "test2"
-                # This will be "test1test2"
-                # So if next line is bigger size then full string, it's the end of multiple string line
-                i = 0
-                line_size += len(lst_line[i_lineno + i].strip())
-                while line_size < str_size + extra_size:
-                    i += 1
-                i_lineno += i + 1
-            return i_lineno
-
-        def _get_recursive_lineno(self, item, set_lineno, lst_line):
-            if hasattr(item, "lineno"):
-                lineno = getattr(item, "lineno")
-                if lineno:
-                    i_lineno = item.lineno
-                    if type(item) is ast.Str:
-                        if "\n" in item.s:
-                            # -1 to ignore last \n
-                            i_lineno = item.lineno - item.s.count("\n")
-                        elif lst_line[i_lineno - 1][-3:] == '"""':
-                            i_lineno = self._get_nb_line_multiple_string(
-                                item, lst_line, i_lineno, extra_size=6
-                            )
-                        elif lst_line[i_lineno - 1][-1] == '"':
-                            i_lineno = self._get_nb_line_multiple_string(
-                                item, lst_line, i_lineno
-                            )
-                    set_lineno.add(i_lineno)
-
-            # Do recursive search, search last line of code
-            lst_attr = [
-                "body",
-                "finalbody",
-                "orelse",
-                "handlers",
-                "test",
-                "right",
-                "left",
-                "value",
-                "exc",
-                "ctx",
-                "func",
-                "args",
-                "elts",
-            ]
-            for attr in lst_attr:
-                if not hasattr(item, attr):
-                    continue
-                lst_attr_item = getattr(item, attr)
-                if not lst_attr_item:
-                    continue
-                if type(lst_attr_item) is list:
-                    for attr_item in lst_attr_item:
-                        if attr_item:
-                            self._get_recursive_lineno(
-                                attr_item, set_lineno, lst_line
-                            )
-                elif type(lst_attr_item) in (
-                    ast.Compare,
-                    ast.Call,
-                    ast.Str,
-                    ast.Attribute,
-                    ast.JoinedStr,
-                    ast.BinOp,
-                    ast.NameConstant,
-                    ast.Name,
-                    ast.arguments,
-                    ast.Load,
-                    ast.List,
-                    ast.IfExp,
-                    ast.Subscript,
-                    ast.UnaryOp,
-                    ast.BoolOp,
-                    ast.Dict,
-                    ast.Tuple,
-                    bool,
-                ):
-                    # Check type, but in fact, can accept all type.
-                    # This check is only to understand what style of code we read
-                    self._get_recursive_lineno(
-                        lst_attr_item, set_lineno, lst_line
-                    )
-                else:
-                    _logger.warning(
-                        f"From get recursive {attr} unknown type"
-                        f" {type(lst_attr_item)}."
-                    )
-
-        def _get_min_max_no_line(self, node, lst_line):
-            # hint node.name == ""
-            set_lineno = set()
-            lst_body = []
-            if len(node.body) > 1:
-                lst_body.append(node.body[0])
-                lst_body.append(node.body[-1])
-            else:
-                lst_body.append(node.body[0])
-            for body in lst_body:
-                self._get_recursive_lineno(body, set_lineno, lst_line)
-            return min(set_lineno), max(set_lineno)
-
-        def search_import(self, lst_line):
-            # get all line until meet "class "
-            i = 0
-            for line in lst_line:
-                if line.startswith("class "):
-                    break
-                i += 1
-            else:
-                _logger.warning(
-                    "Don't know what to do when missing class in python"
-                    " file..."
-                )
-
-            str_code = "\n".join(lst_line[:i])
-            str_code = str_code.strip()
-            if "'''" in str_code:
-                str_code = str_code.replace("'''", "\\'''")
-            if "\\n" in str_code:
-                str_code = str_code.replace("\\n", "\\\\n")
-            d = {
-                "m2o_model": self.model_id.id,
-                "m2o_module": self.module.id,
-                "code": str_code,
-                "name": "header",
-                "is_templated": True,
-            }
-            if not self.module.env["code.generator.model.code.import"].search(
-                [
-                    ("m2o_model", "=", self.model_id.id),
-                    ("m2o_module", "=", self.module.id),
-                    ("code", "=", str_code),
-                ],
-                limit=1,
-            ):
-                self.module.env["code.generator.model.code.import"].create(d)
-
-        def search_method(self, class_model_ast, lst_line, module):
-            sequence = -1
-            lst_body = [a for a in class_model_ast.body]
-            for i in range(len(lst_body)):
-                node = lst_body[i]
-                if i + 1 < len(lst_body):
-                    next_node = lst_body[i + 1]
-                else:
-                    next_node = None
-                if type(node) is ast.Assign:
-                    if node.targets:
-                        if node.targets[0].id == "_description":
-                            value = self._fill_search_field(node.value)
-                            self.model_id.description = value
-                        elif node.targets[0].id == "_inherit":
-                            value = self._fill_search_field(node.value)
-                            if type(value) is list:
-                                model_id = module.env["ir.model"].search(
-                                    [("model", "in", value)]
-                                )
-                            else:
-                                model_id = module.env["ir.model"].search(
-                                    [("model", "=", value)]
-                                )
-                            if not model_id:
-                                _logger.warning(
-                                    f"Cannot identify model {value}."
-                                )
-                            else:
-                                self.model_id.add_model_inherit(model_id)
-                        elif node.targets[0].id == "_sql_constraints":
-                            lst_value = self._fill_search_field(node.value)
-                            constraint_ids = module.env[
-                                "ir.model.constraint"
-                            ].search(
-                                [("module", "=", module.template_module_id.id)]
-                            )
-                            model_name = self.model_id.model.replace(".", "_")
-                            for value in lst_value:
-                                name = value[0]
-                                db_name = f"{model_name}_{name}"
-                                definition = value[1]
-                                message = value[2]
-                                _logger.warning(
-                                    "Ignore next error about ALTER TABLE DROP"
-                                    " CONSTRAINT."
-                                )
-                                constraint_id = constraint_ids.search(
-                                    [("name", "=", db_name)]
-                                )
-                                if constraint_id:
-                                    constraint_id.definition = definition
-                                    constraint_id.message = message
-                elif type(node) is ast.FunctionDef:
-                    sequence += 1
-                    d = {
-                        "m2o_model": self.model_id.id,
-                        "m2o_module": self.module.id,
-                        "name": node.name,
-                        "sequence": sequence,
-                        "is_templated": True,
-                    }
-                    if node.args:
-                        d["param"] = self._extract_argument(node.args)
-                    if node.returns:
-                        d["returns"] = node.returns.id
-                    if node.decorator_list:
-                        str_decorator = self._extract_decorator(
-                            node.decorator_list
-                        )
-                        d["decorator"] = str_decorator
-                    no_line_min, no_line_max = self._get_min_max_no_line(
-                        node, lst_line
-                    )
-                    # Ignore this no_line_max, bug some times.
-                    # no_line_min = min([a.lineno for a in node.body])
-                    if next_node:
-                        no_line_max = next_node.lineno - 1
-                    else:
-                        # TODO this will bug with multiple class
-                        no_line_max = len(lst_line)
-                    codes = ""
-                    for line in lst_line[no_line_min - 1 : no_line_max]:
-                        if line.startswith(" " * 8):
-                            str_line = line[8:]
-                        else:
-                            str_line = line
-                        codes += f"{str_line}\n"
-                    # codes = "\n".join(lst_line[no_line_min - 1:no_line_max])
-                    if "'''" in codes:
-                        codes = codes.replace("'''", "\\'''")
-                    if "\\n" in codes:
-                        codes = codes.replace("\\n", "\\\\n")
-                    d["code"] = codes.strip()
-                    self.module.env["code.generator.model.code"].create(d)
 
     @staticmethod
     def _fmt_underscores(word):
@@ -3664,10 +2393,8 @@ class CodeGeneratorWriter(models.Model):
             for model in lst_model:
                 model = model.strip()
                 if model:
-                    module.view_file_sync[model] = self.ExtractorView(
-                        module, model
-                    )
-                    module.module_file_sync[model] = self.ExtractorModule(
+                    module.view_file_sync[model] = ExtractorView(module, model)
+                    module.module_file_sync[model] = ExtractorModule(
                         module, model, module.view_file_sync[model]
                     )
 
@@ -3794,654 +2521,3 @@ class CodeGeneratorWriter(models.Model):
 
     def get_list_path_file(self):
         return self.list_path_file.split(";")
-
-
-class CodeGeneratorData:
-    def __init__(self, module, path):
-        self._lst_models_init_imports = []
-        self._lst_wizards_init_imports = []
-        self._lst_controllers_init_imports = []
-        self._lst_path_file = set()
-        self._dct_data_depend = defaultdict(list)
-        self._dct_data_metadata_file = defaultdict(list)
-        self._path = path
-        self._module_name = module.name.lower().strip()
-        self._module_path = os.path.join(path, self._module_name)
-        self._data_path = "data"
-        self._demo_path = "demo"
-        self._tests_path = "tests"
-        self._i18n_path = "i18n"
-        self._migrations_path = "migrations"
-        self._readme_path = "readme"
-        self._components_path = "components"
-        self._models_path = "models"
-        self._css_path = os.path.join("static", "src", "scss")
-        self._security_path = "security"
-        self._views_path = "views"
-        self._wizards_path = "wizard"
-        self._controllers_path = "controllers"
-        self._reports_path = "report"
-        self._static_description_path = os.path.join("static", "description")
-        self._lst_manifest_data_files = []
-        self._dct_import_dir = defaultdict(list)
-        self._dct_extra_module_init_path = defaultdict(list)
-        self._dct_view_id = {}
-        # Copy not_supported_files first and permit code to overwrite it
-        self.copy_not_supported_files(module)
-
-    def copy_not_supported_files(self, module):
-        # TODO this is an hack to get code_generator module to search not_supported_files
-        # TODO refactor this and move not_supported_files in models, this is wrong conception
-        if not module.icon:
-            return
-
-        origin_path = os.path.normpath(
-            os.path.join(module.icon, "..", "..", "..", "not_supported_files")
-        )
-        if os.path.isdir(origin_path):
-            for root, dirs, files in os.walk(origin_path):
-                for file in files:
-                    file_path = os.path.join(root, file)
-                    relative_path = os.path.join(
-                        root[len(origin_path) + 1 :], file
-                    )
-                    _, ext = os.path.splitext(relative_path)
-                    is_data_file = ext == ".xml"
-                    self.copy_file(
-                        file_path, relative_path, data_file=is_data_file
-                    )
-
-    @staticmethod
-    def os_make_dirs(path, exist_ok=True):
-        """
-        Util function to wrap the makedirs method
-        :param path:
-        :param exist_ok:
-        :return:
-        """
-        os.makedirs(path, exist_ok=exist_ok)
-
-    @property
-    def lst_path_file(self):
-        return list(self._lst_path_file)
-
-    @property
-    def dct_data_depend(self):
-        return self._dct_data_depend
-
-    @property
-    def dct_data_metadata_file(self):
-        return self._dct_data_metadata_file
-
-    @property
-    def module_path(self):
-        return self._module_path
-
-    @property
-    def data_path(self):
-        return self._data_path
-
-    @property
-    def demo_path(self):
-        return self._demo_path
-
-    @property
-    def tests_path(self):
-        return self._tests_path
-
-    @property
-    def i18n_path(self):
-        return self._i18n_path
-
-    @property
-    def migrations_path(self):
-        return self._migrations_path
-
-    @property
-    def readme_path(self):
-        return self._readme_path_path
-
-    @property
-    def components_path(self):
-        return self._components_path
-
-    @property
-    def models_path(self):
-        return self._models_path
-
-    @property
-    def css_path(self):
-        return self._css_path
-
-    @property
-    def security_path(self):
-        return self._security_path
-
-    @property
-    def views_path(self):
-        return self._views_path
-
-    @property
-    def wizards_path(self):
-        return self._wizards_path
-
-    @property
-    def controllers_path(self):
-        return self._controllers_path
-
-    @property
-    def reports_path(self):
-        return self._reports_path
-
-    @property
-    def static_description_path(self):
-        return self._static_description_path
-
-    @property
-    def dct_view_id(self):
-        return self._dct_view_id
-
-    @property
-    def lst_manifest_data_files(self):
-        return self._lst_manifest_data_files
-
-    @property
-    def lst_import_dir(self):
-        return list(self._dct_import_dir.keys())
-
-    def add_view_id(self, name, str_id):
-        self._dct_view_id[name] = str_id
-
-    def add_module_init_path(self, component, import_line):
-        self._dct_extra_module_init_path[component].append(import_line)
-
-    def _get_lst_files_data_depends(self, lst_meta):
-        set_files = set()
-        for meta in lst_meta:
-            for file_name, lst_key in self.dct_data_metadata_file.items():
-                if meta in lst_key:
-                    set_files.add(file_name)
-                    break
-            else:
-                _logger.error(f"Cannot find key {meta}.")
-        return list(set_files)
-
-    def reorder_manifest_data_files(self):
-        lst_manifest = []
-        dct_hold_file = {}
-        for manifest_data in self._lst_manifest_data_files:
-            if manifest_data in self.dct_data_depend.keys():
-                # find depend and report until order is right
-                lst_meta = self.dct_data_depend.get(manifest_data)
-                lst_files_depends = self._get_lst_files_data_depends(lst_meta)
-                dct_hold_file[manifest_data] = lst_files_depends
-            else:
-                lst_manifest.append(manifest_data)
-
-        i = 0
-        max_i = len(dct_hold_file) + 1
-        while dct_hold_file and i < max_i:
-            i += 1
-            for new_ele, lst_depend in dct_hold_file.items():
-                final_index = -1
-                for depend in lst_depend:
-                    try:
-                        index = lst_manifest.index(depend)
-                    except ValueError:
-                        # element not in list, continue
-                        final_index = -1
-                        break
-                    if index > final_index:
-                        final_index = index
-                if final_index >= 0:
-                    lst_manifest.insert(final_index + 1, new_ele)
-                    del dct_hold_file[new_ele]
-                    # Need to break or crash on loop because dict has change
-                    break
-        if dct_hold_file:
-            _logger.error(f"Cannot reorder all manifest file: {dct_hold_file}")
-        self._lst_manifest_data_files = lst_manifest
-
-    def copy_directory(self, source_directory_path, directory_path):
-        """
-        Copy only directory without manipulation
-        :param source_directory_path:
-        :param directory_path:
-        :return:
-        """
-        absolute_path = os.path.join(
-            self._path, self._module_name, directory_path
-        )
-        # self.check_mkdir_and_create(absolute_path, is_file=False)
-        status = shutil.copytree(source_directory_path, absolute_path)
-
-    def copy_file(
-        self,
-        source_file_path,
-        file_path,
-        data_file=False,
-        search_and_replace=[],
-    ):
-        # TODO if no search_and_replace, use system copy instead of read file and write
-        # TODO problem, we need to add the filename in the system when calling write_file_*
-        # TODO or document it why using this technique
-        with open(source_file_path, "rb") as file_source:
-            content = file_source.read()
-
-        if search_and_replace:
-            # switch binary to string
-            content = content.decode("utf-8")
-            for search, replace in search_and_replace:
-                content = content.replace(search, replace)
-            self.write_file_str(file_path, content, data_file=data_file)
-        else:
-            self.write_file_binary(file_path, content, data_file=data_file)
-
-    def write_file_lst_content(
-        self,
-        file_path,
-        lst_content,
-        data_file=False,
-        insert_first=False,
-        empty_line_end_of_file=True,
-    ):
-        """
-        Function to create a file with some content
-        :param file_path:
-        :param lst_content:
-        :param data_file:
-        :param insert_first:
-        :param empty_line_end_of_file:
-        :return:
-        """
-
-        str_content = "\n".join(lst_content)
-        if empty_line_end_of_file and str_content and str_content[-1] != "\n":
-            str_content += "\n"
-
-        content = str_content.encode("utf-8")
-
-        try:
-            self.write_file_binary(
-                file_path,
-                content,
-                data_file=data_file,
-                insert_first=insert_first,
-            )
-        except Exception as e:
-            _logger.error(e)
-            raise e
-
-    def write_file_str(
-        self, file_path, content, mode="w", data_file=False, insert_first=False
-    ):
-        """
-        Function to create a file with some binary content
-        :param file_path:
-        :param content:
-        :param mode:
-        :param data_file:
-        :param insert_first:
-        :return:
-        """
-        self.write_file_binary(
-            file_path,
-            content,
-            mode=mode,
-            data_file=data_file,
-            insert_first=insert_first,
-        )
-
-    def write_file_binary(
-        self,
-        file_path,
-        content,
-        mode="wb",
-        data_file=False,
-        insert_first=False,
-    ):
-        """
-        Function to create a file with some binary content
-        :param file_path:
-        :param content:
-        :param mode:
-        :param data_file: Will be add in manifest
-        :param insert_first:
-        :return:
-        """
-
-        # file_path suppose to be a relative path
-        if file_path[0] == "/":
-            _logger.warning(f"Path {file_path} not suppose to start with '/'.")
-            file_path = file_path[1:]
-
-        absolute_path = os.path.join(self._path, self._module_name, file_path)
-        self._lst_path_file.add(absolute_path)
-
-        if data_file and file_path not in self._lst_manifest_data_files:
-            if insert_first:
-                self._lst_manifest_data_files.insert(0, file_path)
-            else:
-                self._lst_manifest_data_files.append(file_path)
-
-        self._check_import_python_file(file_path)
-
-        self.check_mkdir_and_create(absolute_path)
-
-        with open(absolute_path, mode) as file:
-            file.write(content)
-
-    @staticmethod
-    def _split_path_all(path):
-        all_parts = []
-        while 1:
-            parts = os.path.split(path)
-            if parts[0] == path:  # sentinel for absolute paths
-                all_parts.insert(0, parts[0])
-                break
-            elif parts[1] == path:  # sentinel for relative paths
-                all_parts.insert(0, parts[1])
-                break
-            else:
-                path = parts[0]
-                all_parts.insert(0, parts[1])
-        return all_parts
-
-    def _check_import_python_file(self, file_path):
-        if file_path and file_path[-3:] == ".py":
-            dir_name = os.path.dirname(file_path)
-            if dir_name == "tests":
-                # Ignore tests python file
-                return
-            if len(self._split_path_all(dir_name)) > 1:
-                # This is a odoo limitation, but we can support it if need it
-                _logger.warning(
-                    "You add python file more depth of 1 directory."
-                )
-                return
-            python_module_name = os.path.splitext(os.path.basename(file_path))[
-                0
-            ]
-            self._dct_import_dir[dir_name].append(python_module_name)
-
-    @staticmethod
-    def check_mkdir_and_create(file_path, is_file=True):
-        if is_file:
-            path_dir = os.path.dirname(file_path)
-        else:
-            path_dir = file_path
-        CodeGeneratorData.os_make_dirs(path_dir)
-
-    def sync_code(self, directory, name):
-        try:
-            # if not os.path.isdir(path_sync_code):
-            #     osmakedirs(path_sync_code)
-            # if module.clean_before_sync_code:
-            path_sync_code = os.path.join(directory, name)
-            if os.path.isdir(path_sync_code):
-                shutil.rmtree(path_sync_code)
-            shutil.copytree(self._module_path, path_sync_code)
-        except Exception as e:
-            _logger.error(e)
-
-    def generate_python_init_file(self, cg_module):
-        for component, lst_module in self._dct_import_dir.items():
-            init_path = os.path.join(component, "__init__.py")
-            if not component:
-                lst_module = [a for a in self._dct_import_dir.keys() if a]
-
-            lst_module.sort()
-
-            cw = CodeWriter()
-
-            if cg_module.license == "AGPL-3":
-                cw.emit(
-                    "# License AGPL-3.0 or later"
-                    " (https://www.gnu.org/licenses/agpl)"
-                )
-                cw.emit()
-            elif cg_module.license == "LGPL-3":
-                cw.emit(
-                    "# License LGPL-3.0 or later"
-                    " (https://www.gnu.org/licenses/lgpl)"
-                )
-                cw.emit()
-            else:
-                _logger.warning(f"License {cg_module.license} not supported.")
-
-            if component:
-                for module in lst_module:
-                    cw.emit(f"from . import {module}")
-            elif lst_module:
-                cw.emit(f"from . import {', '.join(lst_module)}")
-            for extra_import in self._dct_extra_module_init_path.get(
-                component, []
-            ):
-                cw.emit(extra_import)
-            self.write_file_str(init_path, cw.render())
-
-    def flake8_check(self):
-        workspace_path = os.path.normpath(
-            os.path.join(os.path.dirname(__file__), "..", "..", "..", "..")
-        )
-        flake8_bin = os.path.join(workspace_path, ".venv", "bin", "flake8")
-        config_path = os.path.join(workspace_path, ".flake8")
-        cpu_count = os.cpu_count()
-        try:
-            out = subprocess.check_output(
-                [
-                    flake8_bin,
-                    "-j",
-                    str(cpu_count),
-                    f"--config={config_path}",
-                    self.module_path,
-                ]
-            )
-            result = out
-        except subprocess.CalledProcessError as e:
-            result = e.output.decode()
-
-        if result:
-            _logger.warning(result)
-
-    def pylint_check(self):
-        workspace_path = os.path.normpath(
-            os.path.join(os.path.dirname(__file__), "..", "..", "..", "..")
-        )
-        cpu_count = os.cpu_count()
-        try:
-            out = subprocess.check_output(
-                [
-                    f"{workspace_path}/.venv/bin/pylint",
-                    "-j",
-                    str(cpu_count),
-                    "--load-plugins=pylint_odoo",
-                    "-e",
-                    "odoolint",
-                    self.module_path,
-                ]
-            )
-            result = out
-        except subprocess.CalledProcessError as e:
-            result = e.output.decode()
-
-        if result:
-            _logger.warning(result)
-
-    def auto_format(self):
-        workspace_path = os.path.normpath(
-            os.path.join(os.path.dirname(__file__), "..", "..", "..", "..")
-        )
-        max_col = 79
-        use_prettier = True
-        use_format_black = True  # Else, oca-autopep8
-        use_html5print = False
-        enable_xml_formatter = False  # Else, prettier-xml
-        # Manual format with def with programmer style
-        for path_file in self.lst_path_file:
-            relative_path = path_file[len(self.module_path) + 1 :]
-            if path_file.endswith(".py"):
-                # TODO not optimal, too many write for nothing
-                if not use_format_black:
-                    lst_line_write = []
-                    has_change = False
-                    with open(path_file, "r") as source:
-                        for line in source.readlines():
-                            if (
-                                line.lstrip().startswith("def ")
-                                or line.lstrip().startswith("return ")
-                            ) and len(line) > max_col - 1:
-                                has_change = True
-                                next_tab_space = line.find("(") + 1
-                                first_cut = max_col
-                                first_cut = line.rfind(", ", 0, first_cut) + 1
-                                first_part = line[:first_cut]
-                                last_part = line[first_cut:].lstrip()
-                                str_line = (
-                                    f"{first_part}\n{' ' * next_tab_space}{last_part}"
-                                )
-                                lst_line_write.append(str_line[:-1])
-                            else:
-                                lst_line_write.append(line[:-1])
-                    if has_change:
-                        self.write_file_lst_content(
-                            relative_path, lst_line_write
-                        )
-
-            elif path_file.endswith(".js"):
-                if use_prettier:
-                    cmd = f"prettier --write --tab-width 4 {path_file}"
-                    result = subprocess_cmd(cmd)
-                    if result:
-                        _logger.info(result)
-                elif use_html5print:
-                    with open(path_file, "r") as source:
-                        lines = source.read()
-                        try:
-                            lines_out = html5print.JSBeautifier.beautify(
-                                lines, 4
-                            )
-                            self.write_file_str(relative_path, lines_out)
-                        except Exception as e:
-                            _logger.error(e)
-                            _logger.error(f"Check file {path_file}")
-                else:
-                    cmd = (
-                        f"cd {workspace_path};."
-                        f" .venv/bin/activate;css-html-prettify.py {path_file}"
-                    )
-                    result = subprocess_cmd(cmd)
-                    if result:
-                        _logger.warning(result)
-
-            elif path_file.endswith(".scss") or path_file.endswith(".css"):
-                if use_prettier:
-                    cmd = f"prettier --write {path_file}"
-                    result = subprocess_cmd(cmd)
-                    if result:
-                        _logger.info(result)
-                elif use_html5print:
-                    with open(path_file, "r") as source:
-                        lines = source.read()
-                        try:
-                            lines_out = html5print.CSSBeautifier.beautify(
-                                lines, 2
-                            )
-                            self.write_file_str(relative_path, lines_out)
-                        except Exception as e:
-                            _logger.error(e)
-                            _logger.error(f"Check file {path_file}")
-                else:
-                    cmd = (
-                        f"cd {workspace_path};."
-                        f" .venv/bin/activate;css-html-prettify.py {path_file}"
-                    )
-                    result = subprocess_cmd(cmd)
-                    if result:
-                        _logger.warning(result)
-
-            elif path_file.endswith(".html"):
-                if use_prettier:
-                    cmd = f"prettier --write {path_file}"
-                    result = subprocess_cmd(cmd)
-                    if result:
-                        _logger.info(result)
-                elif use_html5print:
-                    with open(path_file, "r") as source:
-                        lines = source.read()
-                        try:
-                            lines_out = html5print.HTMLBeautifier.beautify(
-                                lines, 4
-                            )
-                            self.write_file_str(relative_path, lines_out)
-                        except Exception as e:
-                            _logger.error(e)
-                            _logger.error(f"Check file {path_file}")
-                else:
-                    cmd = (
-                        f"cd {workspace_path};."
-                        f" .venv/bin/activate;css-html-prettify.py {path_file}"
-                    )
-                    result = subprocess_cmd(cmd)
-                    if result:
-                        _logger.warning(result)
-
-            elif path_file.endswith(".xml"):
-                if use_prettier and not enable_xml_formatter:
-                    cmd = (
-                        "prettier --xml-whitespace-sensitivity ignore"
-                        " --prose-wrap always --tab-width 4"
-                        " --no-bracket-spacing --print-width 120 --write"
-                        f" {path_file}"
-                    )
-                    result = subprocess_cmd(cmd)
-                    if result:
-                        _logger.info(result)
-
-        # Automatic format
-        # TODO check diff before and after format to auto improvement of generation
-        if use_format_black:
-            cmd = (
-                f"cd {workspace_path};. .venv/bin/activate;black -l {max_col}"
-                f" --experimental-string-processing -t py37 {self.module_path}"
-            )
-            result = subprocess_cmd(cmd)
-
-            if result:
-                _logger.warning(result)
-        else:
-            maintainer_path = os.path.join(
-                workspace_path, "script", "OCA_maintainer-tools"
-            )
-            cpu_count = os.cpu_count()
-            cmd = (
-                f"cd {maintainer_path};. env/bin/activate;cd"
-                f" {workspace_path};oca-autopep8 -j{cpu_count}"
-                f" --max-line-length {max_col} -ari {self.module_path}"
-            )
-            result = subprocess_cmd(cmd)
-
-            if result:
-                _logger.warning(result)
-
-        if enable_xml_formatter:
-            formatter = xmlformatter.Formatter(
-                indent="4",
-                indent_char=" ",
-                selfclose=True,
-                correct=True,
-                preserve=["pre"],
-                blanks=True,
-            )
-            for path_file in self.lst_path_file:
-                if path_file.endswith(".xml"):
-                    relative_path = path_file[len(self.module_path) + 1 :]
-                    self.write_file_binary(
-                        relative_path, formatter.format_file(path_file)
-                    )
-
-
-def subprocess_cmd(command):
-    process = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
-    proc_stdout = process.communicate()[0].strip()
-    return proc_stdout
